Fix: fixed rename in apps
